fixing aws issue

+++ app/forms/quantity_form.py
@@ -1,7 +1,7 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField, SubmitField
+from wtforms import DecimalField, IntegerField, SubmitField
 from wtforms.validators import data_required
 
 class CreateQuantityForm(FlaskForm):
-    quantity = IntegerField('Quantity', validators=[data_required()])
+    quantity = DecimalField('Quantity', validators=[data_required()])
     submit = SubmitField('Submit')
