+++ app/models/ingredient_recipe.py
@@ -9,7 +9,7 @@ class IngredientRecipe(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     ingredient_id= db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod('ingredients.id')), nullable=False)
     recipe_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod('recipes.id')), nullable=False)
-    quantity = db.Column(db.Integer, nullable=False)
+    quantity = db.Column(db.FLOAT, nullable=False)
 
     ingredients = db.relationship('Ingredient', back_populates='recipe_ingredient')
